requirements.txt
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
grpcio==1.66.1
idna==3.8
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.0.8
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
protobuf==5.28.0
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.1.0
rapidfuzz==3.9.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
tomlkit==0.13.2
trove-classifiers==2024.7.2
urllib3==2.2.2
virtualenv==20.26.3
xattr==1.1.0
